requirements.txt
pyyaml
pillow
path